Using it as sensor
